Defer to service discoery in get_nginxhost_ip.py
This will add a few lines of python to defer to service discovery
in kubernetes. If http://geonode:80 is resolvable from the get-go,
then that will be deferred to and the script will exit. If not, the
rest of the logic which is key to running geonode with compose will
continue to run normally.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>